Fix 4th list item
PROBLEM

The text:

    managed and further developed independently of any single vendor in
    a process open to the equal participation of competitors and third
    parties;

is not enclosed in a list item tag.

SOLUTION

The Open Standard definition we refer to on the updated Open Standards
page is a reference to our own OS definition available here:

    https://fsfe.org/activities/os/def.en.html

where it is clear that the above text has to be the 4th point of the
enclosing orderd list tag.

NOTE

I have fake updated sq translation, just translated, in order to
avoid the warning red (not updatet) box.

svn path=/trunk/; revision=34474