tagtool: remove entire line when removing tags
I've added extended the sed command for removing a tag so that if only
an empty line would remain after removing the tag, the line is removed
entirely.

As a side-note, I've improved the robustness of the sed commands in
general, because previously I had erraneously matched on non-word
characters when I meant to match on spaces.