OpenPGP improvements for people (#50)
Two new fields have been added to person elements:
 * fingerprint
 * keyhref

fingerprint takes a hexadecimal OpenPGP key fingerprint without any
spaces, converts this to an openpgp4fpr URI, and adds a link pointing to
this URI to the person's profile.

keyhref takes a URL to a plain-text, ASCII-armoured OpenPGP public key
and adds a link pointing to the URL to the person's profile.

These fields are not yet explicitly restricted to valid contents.

fingerprints and keyhrefs have been added for all team members whose
keys are stored in the repository or securely linked to from their
profile page. It was not possible to determine Greve's preferred key, so
fingerprint was omitted for Greve.