Replace Upcycle Android videos with embedded PeerTube videos
As discussed on the system-hackers mailinglist, here is the idea to replace the
videos imports from download.fsfe.org. The German translation has the German
video embedded. Non-English pages have the English video with the subtitles of
the corresponding language.

This is just an initial commit, there is more work to be done to ensure the
experience is up to par with the previous experience.

Signed-off-by: Nico Rikken <nico.rikken@fsfe.org>