Updated German translation of freesoftware.de.xhtml (#2366)
WIP:
This is my very first PR at FSFE and git as well. I hope I'm doing it right.
Bonnie gave great support to me in learning git (thanking her very much for her patience and explanations).

The page was marked "outdated" so I did an update to the translation including additions and some minor wordings which I think are more appropriate in their contexts.

The branch in my origin is referred, thus  do not upload any file here. Hoping that this is correct :-)

Co-authored-by: jensmohr <info@jensmohr.de>
Reviewed-on: https://git.fsfe.org/FSFE/fsfe-website/pulls/2366
Reviewed-by: bonnie <bonnie@fsfe.org>
Co-authored-by: Jens Christian Mohr <jotceem@fsfe.org>
Co-committed-by: Jens Christian Mohr <jotceem@fsfe.org>