Danish_translations_of_two_news_corrected (#2610)
Two news items in this commit. The first was translated by mistake (opened the wrong file) but is added here in the hope that the work isn't wasted.

New attepmt at git commit. My first commit on this had the original text commented out, badly - but now it should be clean.

Still hoping someone can proofread formalities (and Danish).

Co-authored-by: Slartibartfast <slartibartfast@fsfe.org>
Reviewed-on: https://git.fsfe.org/FSFE/fsfe-website/pulls/2610
Co-authored-by: Øjvind Fritjof Arnfred <slartibartfast@fsfe.org>
Co-committed-by: Øjvind Fritjof Arnfred <slartibartfast@fsfe.org>