feat/build-improvements (#4516)
This pr re-formats the build scripts for legibility.
Makes some small tweaks, mainly removing options that don't appear to be used and have no impact.
Updates the help file to be more useful and actually represent all arguments.
Adds the ability to specify languages when building. This is mainly useful for speedups, a full rebuild in just English and French takes only ~3 mins as opposed to 40 for a rebuild of all languages.
There may be some small tweaks not mentioned here, see the commit messages for full details.

This pr does not alter the output of a standard build at all, confirmed using `git diff --no-index`.

Co-authored-by: Darragh Elliott <me@delliott.xyz>
Reviewed-on: https://git.fsfe.org/FSFE/fsfe-website/pulls/4516
Co-authored-by: delliott <delliott@fsfe.org>
Co-committed-by: delliott <delliott@fsfe.org>