Implement YH4F videobox (#4773)
Implemented using PeerTube iframes because now you can disable P2P for iframes [1].
The `talks` element is only added to the English page because the talks are in English.

If YH4F talks are carried out in other languages, it should be as simple as creating a `talks.[lang].xml` file with the links and adding the `talks` element to the respective page. From a GDPR perspective, as long as the talks are hosted in the FSFE's PeerTube and the `p2p` option is set to `0`, there shouldn't be any problems.

[1] https://docs.joinpeertube.org/api/embed-player#p2p

Reviewed-on: https://git.fsfe.org/FSFE/fsfe-website/pulls/4773
Reviewed-by: bonnie <bonnie@fsfe.org>
Co-authored-by: Sofía Aritz <sofiaritz@fsfe.org>
Co-committed-by: Sofía Aritz <sofiaritz@fsfe.org>