feat: build by site (#5031)
phase 1 and 2 now by site folder

This allows for us building each site in a different set of languages, wheres before we built every site in every language any site was in.

At the moment this will build each site in every language that there is at least one file of that language in the site.

So if a site has on file in lang AA, it will be built in lang AA. But if it has no files, it will not be built in that lang.

This is a performance enhancement, will do benchmarks later.

Some more complex heuristics for when we do/do not use a language for a file are being discussed in https://git.fsfe.org/FSFE/fsfe-website/issues/4601 .

Co-authored-by: Darragh Elliott <me@delliott.xyz>
Reviewed-on: https://git.fsfe.org/FSFE/fsfe-website/pulls/5031
Reviewed-by: tobiasd <tobiasd@fsfe.org>
Reviewed-by: Sofía Aritz <sofiaritz@fsfe.org>
Co-authored-by: delliott <delliott@fsfe.org>
Co-committed-by: delliott <delliott@fsfe.org>