First try with podcast:transcripts (#5171)
It's my first time with xsl so I would appreciate a review to be sure that I don't break anything.

I assume that `/buildinfo/document/podcast/transcript/url` is processed BEFORE the xls transformations, so the `translate` function can replace the extension `html` to `vtt`.

A better approach would be to not add the extension on each episode file, and code the inclusion on `podcast.rss.xsl`

And the ideal solution would be the above + not duplicating the subtitle file making the `transcript/episode-XX.en.xhtml` "import" the data from the VTT file that is needed by the podcast apps.

Then the VTT file can also be used on the web-player like in https://kdeexpress.gitlab.io/posts/kdeexpress/48-kde-express/#vtt-player

Co-authored-by: David Marzal <4461666-Marzal@users.noreply.gitlab.com>
Reviewed-on: https://git.fsfe.org/FSFE/fsfe-website/pulls/5171
Co-authored-by: David Marzal <dmarzal@fsfe.org>
Co-committed-by: David Marzal <dmarzal@fsfe.org>