refactor: remove one text area, notify that "Custom" needs content
People always forget to fill out the custom field which is strange but kinda understandable

Here I try to improve this by removing an unneeded text field and adding a nice warning in the comment field, that only shows when the comment field is shown.

Requiring the comment field only when shown poorly does not seem to work in plain HTML so we needed to remove it.