ci: push to docs (#5274)
This seems like something we should be doing.

I have some vague memory of trying this before and there being issues as we are coming from the `fsfe` namespace rather than the `fsfe-system-hackers` namespace, but we should be able to get it working eventually regardless.

I will let it run on prs for testing, but before merge we should have it only run on commits to master

Co-authored-by: Darragh Elliott <me@delliott.net>
Reviewed-on: https://git.fsfe.org/FSFE/fsfe-website/pulls/5274