fix: remove unneeded caching from website (#5392)
In https://git.fsfe.org/FSFE/fsfe-website/pulls/5333 I added docker image caching using Gitea artifact Registry.

This turned out to be a waste of time, that actually slowed things down.

This is because the docker instance on the drone runner host actually caches the images internally anyway, so the time cost of downloading an image, rebuilding it and re pushing it was being paid for no reason or benefit.

This pr removes the caching, but does translate a step to ensure that the build docker container can be built in a pr, which is a good check to have pre merge to master.

Co-authored-by: Darragh Elliott <me@delliott.net>
Reviewed-on: https://git.fsfe.org/FSFE/fsfe-website/pulls/5392
Co-authored-by: delliott <delliott@fsfe.org>
Co-committed-by: delliott <delliott@fsfe.org>