feat: pre-commit hook to validate xml|xhtml structures versions are the same (#5432)
Adds a pre-commit hook that checks that `xml` and `xhtml` files of the same version have matched elements and attributes.

The value of elements may differ, as that is translated text.
Some attributes are allowed to have different values, for example image alt text.

This is a continuation of @max.mehl 's work in https://git.fsfe.org/FSFE/fsfe-website/issues/2556, and should hopefully enable closing that issue eventually.

Co-authored-by: Darragh Elliott <me@delliott.net>
Reviewed-on: https://git.fsfe.org/FSFE/fsfe-website/pulls/5432
Reviewed-by: tobiasd <tobiasd@fsfe.org>
Co-authored-by: delliott <delliott@fsfe.org>
Co-committed-by: delliott <delliott@fsfe.org>