feat: reuse compliance for python files (#5440)
Adds reuse compliance checks for python files only, using `reuse lint-file`.

It is managed as a pre-commit hook, and hence ran in ci.

Over time we may add more files to be covered, but I suspect we would use something like `reuse.toml` for setting the licensing for `xml/xhtml/xsl` files, as there are just too many for adding the same text snippet to each to be justifiable.

Co-authored-by: Darragh Elliott <me@delliott.net>
Reviewed-on: https://git.fsfe.org/FSFE/fsfe-website/pulls/5440
Co-authored-by: delliott <delliott@fsfe.org>
Co-committed-by: delliott <delliott@fsfe.org>