fix/news-podcasts (#5400)
Should close https://git.fsfe.org/FSFE/fsfe-website/issues/2471

Move podcasts to a per year structure. This allows for them to be displayed in the archive properly and all that jazz.

This required some enhancements of the subdir script to properly handle more nesting levels, but that is all acceptable.

Also move the transcripts, to a year structure, for consistency.

I have added rewrite rules for the old locations of the podcast episodes, breakage should be minimal.

I am going to keep this as WIP until @bonnie approves it, as it is possible there are breaking issues with this I have failed to notice.

It is also deployed to [test.fsfe.org](https://test.fsfe.org), so feel free to take a look there.

Co-authored-by: Darragh Elliott <me@delliott.net>
Reviewed-on: https://git.fsfe.org/FSFE/fsfe-website/pulls/5400
Reviewed-by: bonnie <bonnie@fsfe.org>
Reviewed-by: tobiasd <tobiasd@fsfe.org>
Co-authored-by: delliott <delliott@fsfe.org>
Co-committed-by: delliott <delliott@fsfe.org>